:construction: Break dashboard stubs out by user type.
Note that the business analyst gets the data owner dashboard, as they are the same.
