+++ web/src/components/dashboards/UserDashboard.vue
@@ -1,0 +1,34 @@
+<template>
+  <h2 class="d-flex flex-column flex-md-row justify-space-between">
+    Dashboard
+    <v-btn
+      class="mt-4 mt-md-0"
+      color="primary"
+      :to="{ name: 'DatasetsPage' }"
+    >
+      View Datasets
+    </v-btn>
+  </h2>
+  <v-row class="mt-6">
+    <v-col
+      cols="12"
+      md="6"
+    >
+      TODO: DatasetSearchCard
+      <v-skeleton-loader
+        type="card"
+        boilerplate
+      />
+    </v-col>
+    <v-col
+      cols="12"
+      md="6"
+    >
+      TODO: MySubscriptionsCard
+      <v-skeleton-loader
+        type="card"
+        boilerplate
+      />
+    </v-col>
+  </v-row>
+</template>
