+++ web/src/pages/DashboardPage.vue
@@ -1,105 +1,41 @@
 <template>
   <v-container>
-    <h2 class="d-flex flex-column flex-md-row justify-space-between">
-      Dashboard
-      <div class="d-flex flex-column flex-md-row mt-4 mt-md-0">
-        <v-btn
-          color="primary"
-          variant="outlined"
-          :to="{ name: 'DatasetsPage' }"
-        >
-          View Datasets
-        </v-btn>
-        <v-btn
-          class="ml-md-4 mt-4 mt-md-0"
-          color="primary"
-          :to="{ name: 'DatasetNewPage' }"
-          >Create Dataset</v-btn
-        >
-      </div>
-    </h2>
-    <v-row class="mt-6">
-      <v-col
-        cols="12"
-        md="6"
-      >
-        TODO: DatasetSearchCard
-        <v-skeleton-loader
-          type="card"
-          boilerplate
-        />
-      </v-col>
-      <v-col
-        cols="12"
-        md="6"
-      >
-        TODO: MySubscriptionsCard
-        <v-skeleton-loader
-          type="card"
-          boilerplate
-        />
-        TODO: MyDataCard
-        <v-skeleton-loader
-          type="card"
-          boilerplate
-        />
-      </v-col>
-    </v-row>
-    <v-row>
-      <v-col>
-        TODO: MyRequestsCard
-        <v-skeleton-loader
-          type="card"
-          boilerplate
-        />
-      </v-col>
-    </v-row>
-    <v-divider
-      thickness="1"
-      class="my-8 border-opacity-100"
-    />
-    <v-row>
-      <v-col
-        cols="12"
-        md="6"
-      >
-        <QaScenariosCard />
-      </v-col>
-      <v-col
-        cols="12"
-        md="6"
-      >
-        <v-btn
-          class="mr-4"
-          color="primary"
-          variant="outlined"
-          :loading="isLoadingUserGroups"
-          @click="syncUserGroups"
-          >Sync User Groups</v-btn
-        >
-      </v-col>
-    </v-row>
+    <component :is="dashboardComponent" />
   </v-container>
 </template>
 
 <script lang="ts" setup>
-import { ref } from "vue"
+import { computed, defineAsyncComponent } from "vue"
 
-import userGroupsApi from "@/api/user-groups-api"
-import { useBreadcrumbs } from "@/use/use-breadcrumbs"
+import useBreadcrumbs from "@/use/use-breadcrumbs"
+import useCurrentUser, { RoleTypes } from "@/use/use-current-user"
 
-import QaScenariosCard from "@/components/qa-scenarios/QaScenariosCard.vue"
+const SystemAdminDashboard = defineAsyncComponent(
+  () => import("@/components/dashboards/SystemAdminDashboard.vue")
+)
+const DataOwnerDashboard = defineAsyncComponent(
+  () => import("@/components/dashboards/DataOwnerDashboard.vue")
+)
+const UserDashboard = defineAsyncComponent(
+  () => import("@/components/dashboards/UserDashboard.vue")
+)
 
-const isLoadingUserGroups = ref(false)
+const { currentUser } = useCurrentUser()
 
-async function syncUserGroups() {
-  isLoadingUserGroups.value = true
-  try {
-    await userGroupsApi.sync()
-  } finally {
-    isLoadingUserGroups.value = false
+const dashboardComponent = computed(() => {
+  if (currentUser.value?.roleTypes.includes(RoleTypes.SYSTEM_ADMIN)) {
+    return SystemAdminDashboard
   }
-}
+
+  if (
+    currentUser.value?.roleTypes.includes(RoleTypes.BUSINESS_ANALYST) ||
+    currentUser.value?.roleTypes.includes(RoleTypes.DATA_OWNER)
+  ) {
+    return DataOwnerDashboard
+  }
+
+  return UserDashboard
+})
 
 const { setBreadcrumbs } = useBreadcrumbs()
